ai-terminal/requirements.txt
# Python zavislosti pro AI Terminal add-on

# Anthropic SDK pro Claude API
anthropic>=0.40.0

# MQTT klient
paho-mqtt>=2.0.0

# YAML zpracovani
PyYAML>=6.0
ruamel.yaml>=0.18.0

# HTTP klient pro HA API
httpx>=0.27.0
aiohttp>=3.9.0

# Utility
python-dateutil>=2.8.0
colorama>=0.4.6
rich>=13.7.0
click>=8.1.0

# JSON schema validace
jsonschema>=4.21.0

# Diff a patch pro YAML soubory
deepdiff>=7.0
